include: move x3_utils into subdir, fix install
