+++ include/depend/depend_parser.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "../x3_utils.hpp"
 #include "depend.hpp"
+#include "misc/x3_utils.hpp"
 
 #include <boost/spirit/home/x3.hpp>
